+++ src/Config.php
@@ -1110,10 +1110,6 @@ class Config extends CommonDBTM
                     $item->showFormLogs();
                     break;
 
-                case 10:
-                    $item->showFormSecurity();
-                    break;
-
                 case 11:
                     Impact::showConfigForm();
                     break;