+++ src/Log.php
@@ -108,12 +108,12 @@ class Log extends CommonDBTM
         ) {
             $nb = countElementsInTable(
                 'glpi_logs',
-                ['itemtype' => $item::class,
+                ['itemtype' => self::getLogItemType($item),
                     'items_id' => $item->getID(),
                 ]
             );
         }
-        return self::createTabEntry(self::getTypeName(1), $nb, $item::class);
+        return self::createTabEntry(self::getTypeName(1), $nb, self::getLogItemType($item));
     }
 
 
@@ -172,7 +172,7 @@ class Log extends CommonDBTM
                         $changes          =  [$id_search_option, $oldval ?? '', $values[$key] ?? ''];
                     }
                 } elseif (
-                    ($val2['linkfield'] == $key && $real_type === $item::class)
+                    ($val2['linkfield'] == $key && $real_type === self::getLogItemType($item))
                        || ($key == $val2['field'] && $val2['table'] == $item->getTable())
                        || ($val2['linkfield'] == $key && $item instanceof Infocom)
                 ) {
@@ -337,7 +337,7 @@ class Log extends CommonDBTM
             return;
         }
 
-        $itemtype = $item::class;
+        $itemtype = self::getLogItemType($item);
         $items_id = $item->getField('id');
 
         $start       = intval(($_GET["start"] ?? 0));
@@ -372,7 +372,7 @@ class Log extends CommonDBTM
             : [],
             'csv_url'           => $CFG_GLPI['root_doc'] . "/front/log/export.php?" . http_build_query([
                 'filter'   => $filters,
-                'itemtype' => $item::class,
+                'itemtype' => $itemtype,
                 'id'       => $item->getId(),
             ]),
         ]);
@@ -399,7 +399,7 @@ class Log extends CommonDBTM
     {
         $DBread = DBConnection::getReadConnection();
 
-        $itemtype  = $item::class;
+        $itemtype  = self::getLogItemType($item);
         $items_id  = $item->getField('id');
         $itemtable = $item->getTable();
 
@@ -955,7 +955,7 @@ class Log extends CommonDBTM
     {
         global $DB;
 
-        $itemtype = $item::class;
+        $itemtype = self::getLogItemType($item);
         $items_id = $item->getField('id');
 
         $iterator = $DB->request([
@@ -996,7 +996,7 @@ class Log extends CommonDBTM
     {
         global $DB;
 
-        $itemtype = $item::class;
+        $itemtype = self::getLogItemType($item);
         $items_id = $item->getField('id');
 
         $affected_fields = ['linked_action', 'itemtype_link', 'id_search_option'];
@@ -1170,7 +1170,7 @@ class Log extends CommonDBTM
     {
         global $DB;
 
-        $itemtype = $item::class;
+        $itemtype = self::getLogItemType($item);
         $items_id = $item->getField('id');
 
         $iterator = $DB->request([
@@ -1483,5 +1483,16 @@ class Log extends CommonDBTM
         }
         $stmt->close();
         static::resetQueue();
+    }
+
+    /**
+     * Check if the given item's class specifies a custom itemtype to use in log entries, and return it. Otherwise, return the class name of the given item.
+     * This is currently used for {@link \Glpi\Security\SecurityConfig} as all historical data is stored for {@link Config}.
+     * @param CommonGLPI $item
+     * @return string
+     */
+    private static function getLogItemtype(CommonGLPI $item): string
+    {
+        return property_exists($item, 'log_itemtype') ? $item::$log_itemtype : $item::class;
     }
 }